refactor(svelte2tsx): transition:/animate: と class:/style: を attributes/ へ切り出す
format_transition_directive* / format_animate_directive_v4 を
attributes/transition.rs、class:/style: ディレクティブ一式を
attributes/class_style.rs へ純粋移動。出力・ロジックの変更なし。
