+++ crates/rsvelte_core/src/svelte2tsx/template/attributes/mod.rs
@@ -1,5 +1,7 @@
 //! Attribute and directive emission.
 
 pub(super) mod action;
+pub(super) mod class_style;
 pub(super) mod let_;
 pub(super) mod svg;
+pub(super) mod transition;
